@@ -15,8 +15,10 @@
     "flowbite": "^2.5.2",
     "flowbite-react": "^0.10.2",
     "framer-motion": "^11.18.0",
+    "gsap": "^3.13.0",
     "highlight.js": "^11.11.1",
-    "next": "15.1.4",
+    "next": "^15.3.1",
+    "prisma": "^6.7.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-icons": "^5.4.0",
