@@ -1,21 +1,67 @@
-// import React from 'react';
+"use client";
+
+import { motion } from "framer-motion";
+import { useEffect, useState } from "react";
+import { FaSearch } from "react-icons/fa";
 
 function Welcome() {
-    return (
-      <div className="text-black dark:text-white py-7">
-        <div className="container mx-auto px-10">
-          <div className="bg-gradient-to-r from-gray-200 to-white dark:from-gray-800 dark:to-gray-900 bg-opacity-20 p-5 rounded-lg shadow-lg transition-opacity duration-1000 ease-in-out opacity-0 animate-fadeIn">
-            <h1 className="text-4xl font-bold mb-4 text-center transition-transform duration-1000 ease-in-out transform translate-y-10 animate-slideIn">
-              Selamat Datang di Tech Talks Blog
-            </h1>
-            <p className="text-lg text-center transition-transform duration-1000 ease-in-out transform translate-y-10 delay-200 animate-slideIn">
-              Platform di mana saya berbagi pemikiran dan mengedukasi semua orang seputaran dunia IT dan bahkan mengenai programming / coding.
-            </p>
-          </div>
-        </div>
+  const [isVisible, setIsVisible] = useState(false);
+
+  useEffect(() => {
+    setIsVisible(true);
+  }, []);
+
+  return (
+    <div className="relative py-16 overflow-hidden  from-white to-gray-100 dark:from-gray-900 dark:to-gray-800">
+      <div className="container relative z-10 mx-auto px-6 sm:px-10">
+        <motion.div
+          initial={{ opacity: 0, y: 20 }}
+          animate={{ opacity: isVisible ? 1 : 0, y: isVisible ? 0 : 20 }}
+          transition={{ duration: 0.8 }}
+          className="bg-white/80 dark:bg-gray-800/80 backdrop-blur-lg p-8 sm:p-10 rounded-2xl shadow-xl border border-gray-100 dark:border-gray-700"
+        >
+          <motion.h1
+            initial={{ opacity: 0, y: 20 }}
+            animate={{ opacity: isVisible ? 1 : 0, y: isVisible ? 0 : 20 }}
+            transition={{ duration: 0.8, delay: 0.2 }}
+            className="text-3xl sm:text-4xl md:text-5xl font-extrabold mb-6 text-center bg-gradient-to-r from-blue-600 to-purple-600 bg-clip-text text-transparent"
+          >
+            Selamat Datang di Tech Talks Blog
+          </motion.h1>
+
+          <motion.p
+            initial={{ opacity: 0, y: 20 }}
+            animate={{ opacity: isVisible ? 1 : 0, y: isVisible ? 0 : 20 }}
+            transition={{ duration: 0.8, delay: 0.4 }}
+            className="text-lg md:text-xl text-center text-gray-600 dark:text-gray-300 mb-8 max-w-3xl mx-auto leading-relaxed"
+          >
+            Platform di mana saya berbagi pemikiran dan mengedukasi semua orang
+            seputaran dunia IT dan bahkan mengenai programming / coding.
+          </motion.p>
+
+          <motion.div
+            initial={{ opacity: 0, scale: 0.9 }}
+            animate={{ opacity: isVisible ? 1 : 0, scale: isVisible ? 1 : 0.9 }}
+            transition={{ duration: 0.5, delay: 0.6 }}
+            className="flex flex-col sm:flex-row justify-center gap-4"
+          >
+            <a
+              href="#latest-posts"
+              className="px-6 py-3 bg-gradient-to-r from-blue-500 to-blue-700 hover:from-blue-600 hover:to-blue-800 text-white font-medium rounded-lg shadow-lg hover:shadow-xl transition-all duration-300 flex items-center justify-center"
+            >
+              Jelajahi Artikel
+            </a>
+            <a
+              href="#search"
+              className="px-6 py-3 bg-white dark:bg-gray-700 text-gray-800 dark:text-white border border-gray-200 dark:border-gray-600 font-medium rounded-lg shadow-md hover:shadow-lg transition-all duration-300 flex items-center justify-center"
+            >
+              <FaSearch className="mr-2" /> Cari Topik
+            </a>
+          </motion.div>
+        </motion.div>
       </div>
-    );
-  }
-  
-  export default Welcome;
-  
+    </div>
+  );
+}
+
+export default Welcome;
